Refs #23128: Review - Separate each builtin annotation implementation
